[Github #672] improvements (#675)
* [gh-#672] improve error response if the connection isn't found

* [gh-#672] show the provider config key

* [gh-#672] update docs for Base-Url-Override

* [gh-#672] set process directly to assume cloud

•

* [gh-#672] better cli example that uses github and cleaner output of results

* [gh-#672] add retries to the activity log

* [gh-#672] sync commands go to the activity log

* [gh-#672] dev: restart server and worker on .env change

* [gh-#672] sync command to activity log

* [gh-#672] schedule update goes to error

* [gh-#672] prettier batch send log if dry run

* [gh-#672] sync command to activity log

* [gh-#672] docs update to await batchSend

* [gh-#672] send models instead of a single model each time

* [gh-#672] error handling

* [gh-#672] updat ejs file

* [gh-#672] working example for cli integration template

* [gh-#672] verify the batchSend sync model matches what is in the sync job

* [gh-#672] better end report based on the added results

* [gh-#672] better unknown provider config error report

* [gh-#672] use payload in the error response

* [gh-#672] don't run as background

* [gh-#672] 0.21.4

* [gh-#672] webhook update to return the job updated at timestamp

* [gh-#672] create models file if it doesn't exist on compile

• add version command

* [gh-#672] add upgrade logic and improve docker down via control c

* [gh-#672] don't deploy if no integration files and add checks for necessary files

* [gh-#672] warn await calls for integration files

* [gh-#672] useEffect for default scope

* [gh-#672] 0.21.5

* [gh-#672] check types

* [gh-#672] 0.21.9

+++ packages/cli/lib/sync.ts
@@ -1,16 +1,25 @@
 import fs from 'fs';
-import path from 'path';
+import path, { dirname } from 'path';
+import { fileURLToPath } from 'url';
 import yaml from 'js-yaml';
 import chalk from 'chalk';
 import chokidar from 'chokidar';
 import * as tsNode from 'ts-node';
 import glob from 'glob';
+import ejs from 'ejs';
 import * as dotenv from 'dotenv';
 import { spawn } from 'child_process';
+import parser from '@babel/parser';
+import traverse, { NodePath } from '@babel/traverse';
+import type { ChildProcess } from 'node:child_process';
+import type * as t from '@babel/types';
 
-import type { NangoConfig, Connection as NangoConnection } from '@nangohq/shared';
+import type { NangoConfig, Connection as NangoConnection, NangoIntegration, NangoIntegrationData } from '@nangohq/shared';
 import { cloudHost, stagingHost, SyncType, syncRunService, nangoConfigFile } from '@nangohq/shared';
 import { hostport, getConnection, NANGO_INTEGRATIONS_LOCATION, buildInterfaces } from './utils.js';
+
+const __filename = fileURLToPath(import.meta.url);
+const __dirname = dirname(__filename);
 
 dotenv.config();
 
@@ -22,43 +31,76 @@ interface RunArgs {
     useServerLastSyncDate?: boolean;
 }
 
+const exampleSyncName = 'github-issue-example';
+
+const createModelFile = (notify = false) => {
+    if (!fs.existsSync(`${NANGO_INTEGRATIONS_LOCATION}/models.ts`)) {
+        const cwd = process.cwd();
+        const configContents = fs.readFileSync(path.resolve(cwd, `${NANGO_INTEGRATIONS_LOCATION}/${nangoConfigFile}`), 'utf8');
+        const configData: NangoConfig = yaml.load(configContents) as unknown as NangoConfig;
+        const { models } = configData;
+        const interfaceDefinitions = buildInterfaces(models);
+        fs.writeFileSync(`${NANGO_INTEGRATIONS_LOCATION}/models.ts`, interfaceDefinitions.join('\n'));
+
+        if (notify) {
+            console.log(
+                chalk.green(
+                    `${NANGO_INTEGRATIONS_LOCATION}/${nangoConfigFile} was updated. The interface file (${NANGO_INTEGRATIONS_LOCATION}/models.ts) was updated to reflect the updated config`
+                )
+            );
+        }
+    }
+};
+
+export const version = () => {
+    const packageJson = JSON.parse(fs.readFileSync(path.resolve(__dirname, '../package.json'), 'utf8'));
+    const dockerComposeYaml = fs.readFileSync(path.resolve(__dirname, '../docker/docker-compose.yaml'), 'utf8');
+    const dockerCompose = yaml.load(dockerComposeYaml) as any;
+
+    const nangoServerImage = dockerCompose.services['nango-server'].image;
+    const nangoWorkerImage = dockerCompose.services['nango-worker'].image;
+
+    const nangoServerVersion = nangoServerImage.split(':').pop();
+    const nangoWorkerVersion = nangoWorkerImage.split(':').pop();
+
+    console.log(chalk.green('Nango Server version:'), nangoServerVersion);
+    console.log(chalk.green('Nango Worker version:'), nangoWorkerVersion);
+    console.log(chalk.green('Nango CLI version:'), packageJson.version);
+};
+
 export const init = () => {
     const data: NangoConfig = {
         integrations: {
-            'github-prod': {
-                'github-users': {
-                    runs: 'every hour',
-                    returns: ['users']
-                },
-                'github-issues': {
+            github: {
+                [exampleSyncName]: {
                     runs: 'every half hour',
-                    returns: ['issues']
+                    returns: ['GithubIssue']
                 }
             },
             'asana-dev': {
                 'asana-projects': {
                     runs: 'every hour',
-                    returns: ['projects']
+                    returns: ['AsanaProject']
                 }
             }
         },
         models: {
-            issues: {
+            GithubIssue: {
                 id: 'integer',
+                owner: 'string',
+                repo: 'string',
+                issue_number: 'number',
                 title: 'string',
-                description: 'string',
-                status: 'string',
-                author: {
-                    avatar_url: 'string'
-                }
+                author: 'string',
+                author_id: 'string',
+                state: 'string',
+                date_created: 'date',
+                date_last_modified: 'date',
+                body: 'string'
             },
-            projects: {
+            AsanaProject: {
                 id: 'number',
                 type: 'string'
-            },
-            users: {
-                id: 'number',
-                name: 'string'
             }
         }
     };
@@ -67,9 +109,75 @@ export const init = () => {
     if (!fs.existsSync(NANGO_INTEGRATIONS_LOCATION)) {
         fs.mkdirSync(NANGO_INTEGRATIONS_LOCATION);
     }
-    fs.writeFileSync(`${NANGO_INTEGRATIONS_LOCATION}/${nangoConfigFile}`, yamlData);
 
-    console.log(chalk.green(`${nangoConfigFile} file has been created`));
+    if (!fs.existsSync(`${NANGO_INTEGRATIONS_LOCATION}/${nangoConfigFile}`)) {
+        fs.writeFileSync(`${NANGO_INTEGRATIONS_LOCATION}/${nangoConfigFile}`, yamlData);
+    }
+
+    // check if a .env file exists and if not create it with some default content
+    if (!fs.existsSync('.env')) {
+        fs.writeFileSync(
+            '.env',
+            `#NANGO_HOSTPORT=https://api-staging.nango.dev
+#NANGO_SECRET_KEY=xxxx-xxx-xxxx
+#NANGO_INTEGRATIONS_LOCATION=use-this-to-override-where-the-nango-integrations-directory-goes
+#NANGO_PORT=use-this-to-override-the-default-3003
+#NANGO_DB_PORT=use-this-to-override-the-default-5432`
+        );
+    }
+
+    console.log(chalk.green(`Nango integrations initialized!`));
+};
+
+export const generate = async () => {
+    const templateContents = fs.readFileSync(path.resolve(__dirname, './integration.ejs'), 'utf8');
+    const githubExampleTemplateContents = fs.readFileSync(path.resolve(__dirname, './integration.github.ejs'), 'utf8');
+
+    const cwd = process.cwd();
+    const configContents = fs.readFileSync(path.resolve(cwd, `${NANGO_INTEGRATIONS_LOCATION}/${nangoConfigFile}`), 'utf8');
+    const configData: NangoConfig = yaml.load(configContents) as unknown as NangoConfig;
+    const { integrations } = configData;
+    const { models } = configData;
+
+    const interfaceDefinitions = buildInterfaces(models);
+
+    fs.writeFileSync(`${NANGO_INTEGRATIONS_LOCATION}/models.ts`, interfaceDefinitions.join('\n'));
+
+    for (let i = 0; i < Object.keys(integrations).length; i++) {
+        const providerConfigKey = Object.keys(integrations)[i] as string;
+        const syncObject = integrations[providerConfigKey] as unknown as { [key: string]: NangoIntegration };
+        const syncNames = Object.keys(syncObject);
+        for (let k = 0; k < syncNames.length; k++) {
+            const syncName = syncNames[k] as string;
+            const syncData = syncObject[syncName] as unknown as NangoIntegrationData;
+            const { returns: models } = syncData;
+            const syncNameCamel = syncName
+                .split('-')
+                .map((word) => word.charAt(0).toUpperCase() + word.slice(1))
+                .join('');
+            const ejsTeamplateContents = syncName === exampleSyncName ? githubExampleTemplateContents : templateContents;
+            const rendered = ejs.render(ejsTeamplateContents, {
+                syncName: syncNameCamel,
+                interfaceNames: models.map((model) => {
+                    const singularModel = model?.charAt(model.length - 1) === 's' ? model.slice(0, -1) : model;
+                    return `${singularModel.charAt(0).toUpperCase()}${singularModel.slice(1)}`;
+                }),
+                mappings: models.map((model) => {
+                    const singularModel = model.charAt(model.length - 1) === 's' ? model.slice(0, -1) : model;
+                    return {
+                        name: model,
+                        type: `${singularModel.charAt(0).toUpperCase()}${singularModel.slice(1)}`
+                    };
+                })
+            });
+
+            if (!fs.existsSync(`${NANGO_INTEGRATIONS_LOCATION}/${syncName}.ts`)) {
+                fs.writeFileSync(`${NANGO_INTEGRATIONS_LOCATION}/${syncName}.ts`, rendered);
+            }
+        }
+    }
+
+    console.log(chalk.green(`Integration files have been created`));
 };
 
 export const run = async (args: string[], options: RunArgs) => {
@@ -133,7 +241,7 @@ export const run = async (args: string[], options: RunArgs) => {
 
     try {
         const results = await syncRun.run(lastSyncDate, true);
-        console.log(results);
+        console.log(JSON.stringify(results, null, 2));
         process.exit(0);
     } catch (e) {
         process.exit(1);
@@ -150,6 +258,10 @@ export const tsc = () => {
         fs.mkdirSync(distDir);
     }
 
+    if (!fs.existsSync(`${NANGO_INTEGRATIONS_LOCATION}/models.ts`)) {
+        createModelFile();
+    }
+
     const rawNangoIntegrationLocation = NANGO_INTEGRATIONS_LOCATION.replace('./', '');
 
     const compiler = tsNode.create({
@@ -159,6 +271,9 @@ export const tsc = () => {
     const integrationFiles = glob.sync(path.resolve(cwd, `${NANGO_INTEGRATIONS_LOCATION}/*.ts`));
     for (const filePath of integrationFiles) {
         try {
+            if (!nangoCallsAreAwaited(filePath)) {
+                return;
+            }
             const result = compiler.compile(fs.readFileSync(filePath, 'utf8'), filePath);
             const jsFilePath = path.join(path.dirname(filePath), path.basename(filePath, '.ts') + '.js');
             const distJSFilePath = jsFilePath.replace(rawNangoIntegrationLocation, `${rawNangoIntegrationLocation}/dist`);
@@ -170,6 +285,36 @@ export const tsc = () => {
             console.error(error);
         }
     }
+};
+
+const nangoCallsAreAwaited = (filePath: string): boolean => {
+    const code = fs.readFileSync(filePath, 'utf-8');
+    let areAwaited = true;
+
+    const ast = parser.parse(code, { sourceType: 'module', plugins: ['typescript'] });
+
+    const message = (call: string, lineNumber: number) =>
+        console.log(chalk.red(`nango.${call}() calls must be awaited in "${filePath}:${lineNumber}". Not awaiting can lead to unexpected results.`));
+
+    const nangoCalls = ['batchSend', 'log', 'getFieldMapping', 'setFieldMapping', 'get', 'post', 'put', 'patch', 'delete'];
+
+    // @ts-ignore
+    traverse.default(ast, {
+        CallExpression(path: NodePath<t.CallExpression>) {
+            const lineNumber = path.node.loc?.start.line as number;
+            const callee = path.node.callee as t.MemberExpression;
+            if (callee.object?.type === 'Identifier' && callee.object.name === 'nango' && callee.property?.type === 'Identifier') {
+                if (path.parent.type !== 'AwaitExpression') {
+                    if (nangoCalls.includes(callee.property.name)) {
+                        message(callee.property.name, lineNumber);
+                        areAwaited = false;
+                    }
+                }
+            }
+        }
+    });
+
+    return areAwaited;
 };
 
 export const tscWatch = () => {
@@ -190,6 +335,10 @@ export const tscWatch = () => {
 
     if (!fs.existsSync(distDir)) {
         fs.mkdirSync(distDir);
+    }
+
+    if (!fs.existsSync(`${NANGO_INTEGRATIONS_LOCATION}/models.ts`)) {
+        createModelFile();
     }
 
     watcher.on('add', (filePath: string) => {
@@ -223,12 +372,14 @@ export const tscWatch = () => {
         });
 
         try {
+            if (!nangoCallsAreAwaited(filePath)) {
+                return;
+            }
             const result = compiler.compile(fs.readFileSync(filePath, 'utf8'), filePath);
             const jsFilePath = path.join(path.dirname(filePath), path.basename(filePath, '.ts') + '.js');
 
             const distJSFilePath = jsFilePath.replace(rawNangoIntegrationLocation, `${rawNangoIntegrationLocation}/dist`);
             fs.writeFileSync(distJSFilePath, result);
-
             console.log(chalk.green(`Compiled ${filePath} successfully`));
         } catch (error) {
             console.error(`Error compiling ${filePath}:`);
@@ -242,38 +393,52 @@ export const configWatch = () => {
     const watchPath = `${NANGO_INTEGRATIONS_LOCATION}/${nangoConfigFile}`;
     const watcher = chokidar.watch(watchPath, { ignoreInitial: true });
 
-    watcher.on('add', (filePath) => {
-        buildInterface(filePath);
+    watcher.on('add', () => {
+        createModelFile(true);
     });
 
-    watcher.on('change', (filePath) => {
-        buildInterface(filePath);
+    watcher.on('change', () => {
+        createModelFile(true);
     });
-
-    function buildInterface(filePath: string) {
-        const cwd = process.cwd();
-        const configContents = fs.readFileSync(path.resolve(cwd, `${NANGO_INTEGRATIONS_LOCATION}/${nangoConfigFile}`), 'utf8');
-        const configData: NangoConfig = yaml.load(configContents) as unknown as NangoConfig;
-        const { models } = configData;
-        const interfaces = buildInterfaces(models);
-        fs.writeFileSync(`${NANGO_INTEGRATIONS_LOCATION}/models.ts`, interfaces.join('\n'));
-        console.log(
-            chalk.green(`${filePath} was updated. The interface file (${NANGO_INTEGRATIONS_LOCATION}/models.ts) was updated to reflect the updated config`)
-        );
-    }
 };
+
+let child: ChildProcess | undefined;
+process.on('SIGINT', () => {
+    if (child) {
+        const dockerDown = spawn('docker', ['compose', '-f', 'node_modules/nango/docker/docker-compose.yaml', '--project-directory', '.', 'down'], {
+            stdio: 'inherit'
+        });
+        dockerDown.on('exit', () => {
+            process.exit();
+        });
+    } else {
+        process.exit();
+    }
+});
 
 /**
  * Docker Run
  * @desc spawn a child process to run the docker compose located in the cli
  * Look into https://www.npmjs.com/package/docker-compose to avoid dependency maybe?
  */
-export const dockerRun = () => {
+export const dockerRun = async () => {
     const cwd = process.cwd();
 
-    spawn('docker', ['compose', '-f', 'node_modules/nango/docker/docker-compose.yaml', '--project-directory', '.', 'up', '--build'], {
+    child = spawn('docker', ['compose', '-f', 'node_modules/nango/docker/docker-compose.yaml', '--project-directory', '.', 'up', '--build'], {
         cwd,
         detached: false,
         stdio: 'inherit'
+    });
+
+    await new Promise((resolve, reject) => {
+        child?.on('exit', (code) => {
+            if (code !== 0) {
+                reject(new Error(`Error with the nango docker containers, please check your containers using 'docker ps''`));
+                return;
+            }
+            resolve(true);
+        });
+
+        child?.on('error', reject);
     });
 };
